requirements.txt
numpy
pandas
scikit-learn
python-dotenv
seaborn
mlflow
catboost
xgboost